backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
django-colorfield==0.7.2
django-extra-fields==3.0.2
psycopg2==2.9.3
